format && replact uint32_t by uint64_t

+++ upmem/src/dpu/dpu_embedding.c
@@ -30,27 +30,28 @@ __dma_aligned int32_t tmp_results[MAX_NR_BATCHES];
 
 int
 main() {
-    if(me()==0){
-        mram_read(&input_lengths, &lengths, ALIGN(sizeof(struct query_len),8));
-        indices_len=lengths.indices_len;
-        nr_batches=lengths.nr_batches;
+    if (me() == 0) {
+        mram_read(&input_lengths, &lengths, ALIGN(sizeof(struct query_len), 8));
+        indices_len = lengths.indices_len;
+        nr_batches = lengths.nr_batches;
 
-        uint32_t copied_indices=0;
-        while(copied_indices<indices_len){
-            mram_read(&input_indices[copied_indices],&indices[copied_indices],
-            ALIGN(MIN(2048, (indices_len-copied_indices)*sizeof(uint32_t)),8));
-            copied_indices+=2048/sizeof(uint32_t);
+        uint64_t copied_indices = 0;
+        while (copied_indices < indices_len) {
+            mram_read(&input_indices[copied_indices], &indices[copied_indices],
+                      ALIGN(MIN(2048, (indices_len - copied_indices) * sizeof(uint32_t)), 8));
+            copied_indices += 2048 / sizeof(uint32_t);
         }
 
-        mram_read(input_offsets,offsets,ALIGN(nr_batches*sizeof(uint32_t),8));
+        mram_read(input_offsets, offsets, ALIGN(nr_batches * sizeof(uint32_t), 8));
     }
     barrier_wait(&my_barrier);
 
     for (uint64_t i = me(); i < nr_batches; i += NR_TASKLETS) {
         tmp_results[i] = 0;
-        uint32_t upper_bound = i == nr_batches - 1 ? indices_len : offsets[i + 1];
-        for (uint32_t indices_ptr = offsets[i]; indices_ptr < upper_bound; indices_ptr++) {
-            uint32_t ind = indices[indices_ptr];
+        uint64_t upper_bound = i == nr_batches - 1 ? indices_len : offsets[i + 1];
+        for (uint64_t indices_ptr = offsets[i]; indices_ptr < upper_bound; indices_ptr++) {
+            uint64_t ind = indices[indices_ptr];
+            // printf("ind %lu indices ptr %lu\n", ind, indices_ptr);
             tmp_results[i] += emb_data[ind];
         }
     }
